add live/common.lua file
